Programmer is a experimental game-making interpreted language build on SFML
Programmer is a interpreted language that was build solely for the
purpose of learning to make the languages.

It is meant to be a simple language that can create games with only
logic and graphics.

Example code is included.

It uses SFML to handle graphics.

It is nowhere near complete, but one can make "games" in it at this
stage.